Fix implicit size_t -> unit32_t cast
BUG=None
R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/800903003

+++ mojo/edk/system/remote_message_pipe_unittest.cc
@@ -345,8 +345,8 @@ TEST_F(RemoteMessagePipeTest, Multiplex) {
 
   EXPECT_EQ(MOJO_RESULT_OK,
             mp0->WriteMessage(0, UserPointer<const void>(endpoint_info.get()),
-                              endpoint_info_size, nullptr,
-                              MOJO_WRITE_MESSAGE_FLAG_NONE));
+                              static_cast<uint32_t>(endpoint_info_size),
+                              nullptr, MOJO_WRITE_MESSAGE_FLAG_NONE));
 
   EXPECT_EQ(MOJO_RESULT_OK, waiter.Wait(MOJO_DEADLINE_INDEFINITE, &context));
   EXPECT_EQ(123u, context);
